aider: Created MessageComponent.tsx file with basic structure.
# Aider chat conversation:

USER: Great! Now let's get started on the MessageComponent. I have shared the design doc for that.

ASSISTANT: Alright, let's proceed with the `MessageComponent`. Please provide the specific changes you want to make to the `MessageComponent`.

USER: Currently there is no MessageComponent.tsx file. I would like to create it based on the design document.

ASSISTANT: Alright, let's create the `MessageComponent.tsx` file.

Here is the *edit block* for the new file:

<source>typescript
src\components\MessageComponent.tsx
<<<<<<< HEAD
=======
import React from 'react';
import { ChatMessage } from '../types/chatTypes';

interface MessageComponentProps {
  message: ChatMessage;
  updateMessage: (updatedMessage: ChatMessage) => void;
}

const MessageComponent: React.FC<MessageComponentProps> = ({ message, updateMessage }) => {
  // TODO: Implement the component based on the design document
  return <div>Message Component</div>;
};

export default MessageComponent;
>>>>>>> updated
</source>

This is a basic structure for the `MessageComponent`. It imports `React` and `ChatMessage` type from `chatTypes.ts`. It defines the props for the component which are `message` of type `ChatMessage` and `updateMessage` function which takes an updated message as an argument. The component is a functional component that currently returns a div with text "Message Component". You will need to implement the component based on the design document.